Accounts Page
Here is some progress on the Accounts page. The main page features two cards(Checking and Savings) as well as two buttons to edit the accounts and open a new account. Clicking on "Go to Checking Account" or "Go to Savings Account" will open separate pages. When clicked on, both buttons (Edit Accounts and Open an account) opens a pop-up window. The details in actually opening an account have not been set up yet.